Bunch of bug fixes

+++ src/lib/components/accountComponents/ArchiveAccountPopup.svelte
@@ -3,14 +3,18 @@
   import Button from '$lib/components/sharedComponents/Button.svelte'
   import { openPopup } from '$lib/store'
   import { AccountHierarchy } from '../../../params/accountHierarchy'
-  import { invalidate } from '$app/navigation'
+  import { goto, invalidate } from '$app/navigation'
   import Alert from '$lib/components/sharedComponents/Alert.svelte'
+  import LoadingSpinner from '$lib/components/sharedComponents/LoadingSpinner.svelte'
 
   export let accountId: number
   export let isParent: boolean
   let error: string
+  let isLoading = false
 
   async function onClick () {
+    if (isLoading) return
+    isLoading = true
     const endpoint = isParent ? AccountHierarchy.PARENT_ACCOUNT : AccountHierarchy.ACCOUNT
     const res = await fetch(`/${endpoint}/${accountId}`, {
       method: 'DELETE',
@@ -21,14 +25,17 @@
     if (res.status === 204) {
       $openPopup = false
       await invalidate('data:accounts')
+      await goto('/')
     } else {
       const body = await res.json()
       error = body.error ?? body.message
     }
+    isLoading = false
   }
   function onExit () {
     error = ''
     $openPopup = false
+    isLoading = false
   }
 </script>
 
@@ -39,7 +46,7 @@
     <Alert>{error}</Alert>
   {/if}
   <div>
-    <Button on:click={onClick}>Yes</Button>
+    <Button disabled={isLoading} on:click={onClick}>Yes{#if isLoading}<LoadingSpinner/>{/if}</Button>
     <Button secondary={true} on:click={onExit}>No</Button>
   </div>
 </Popup>
